Refactored move-date module

+++ main/ContentGeneration/move_date.py
@@ -1,9 +1,7 @@
 
-from django.http import HttpResponse, JsonResponse
-from pathlib import Path
+from django.http import JsonResponse
 from django.forms.utils import ErrorDict
 from django.forms import model_to_dict
-from django.shortcuts import redirect
 
 import main.models as models
 import main.forms as forms
@@ -11,8 +9,8 @@ import main.forms as forms
 from main.ContentGeneration.request_forms import DateMoveForm
 from main.Helpers.date_helpers import getValidDateFromSlug, convertDateToUrlTuple
 from main.ContentGeneration.delete_entry import moveImagesOutOfADeleteFolder
-from main.ContentGeneration.content_models import CONTENT_MODELS
-from main.ContentGeneration.content_forms import CONTENT_FORMS
+from main.ContentGeneration.content_factory_models import CONTENT_MODELS
+from main.ContentGeneration.content_factory_update import CONTENT_UPDATE_DATE
 
 
 def checkMoveRequest(post_data):
@@ -24,12 +22,12 @@ def checkMoveRequest(post_data):
     return date_move_form.cleaned_data, None
 
 
-def updateEntryDate(source_slug, destination_slug):
+def createNewEntry(source_entry, destination_slug):
     errors = ErrorDict()
-    entry = models.Entry.objects.get(name=source_slug)
-    moveImagesOutOfADeleteFolder(entry)
+    new_entry = None
+    moveImagesOutOfADeleteFolder(source_entry)
 
-    entry_dict = model_to_dict(entry)
+    entry_dict = model_to_dict(source_entry)
     entry_dict['name'] = destination_slug
     entry_dict['date'] = getValidDateFromSlug(destination_slug)
     entry_dict['content'] = []
@@ -37,50 +35,72 @@ def updateEntryDate(source_slug, destination_slug):
     new_entry_form = forms.EntryForm(entry_dict)
     if new_entry_form.is_valid():
         new_entry_form.save(commit=True)
+        new_entry = models.Entry.objects.get(name=destination_slug)
     else:
         errors["entry"] = new_entry_form.errors
-        return None, errors
+    
+    return new_entry, errors
 
+
+def generateNewContentObjectFromSource(content, new_slug, errors):
+    content_type = content.content_type
+    
+    Model = CONTENT_MODELS[content_type]
+    obj = Model.objects.get(id=content.content_id)
+    new_obj_form = CONTENT_UPDATE_DATE[content_type](obj, new_slug)
+
+    new_instance_id = None
+    if new_obj_form.is_valid():
+        new_obj_form.save(commit=True)
+        new_instance_id = new_obj_form.instance.pk
+        obj.delete()
+    else:
+        errors[f'{content_type}-{obj.id}'] = new_obj_form.errors
+        
+    return new_instance_id
+
+
+def generateNewContentRecordFromSource(content, new_instance_id, errors):
+    new_content_form = forms.ContentForm({
+        'content_type': content.content_type,
+        'content_id': new_instance_id
+    })
+
+    new_content_record_id = None
+    if new_content_form.is_valid():
+        new_content_form.save(commit=True)
+        new_content_record_id = new_content_form.instance.pk
+        content.delete()
+    else:
+        errors[f'content-{content.id}'] = new_content_form.errors
+        
+    return new_content_record_id
+
+
+def updateNewContentIds(content, destination_slug, content_ids, errors):
+    new_instance_id = generateNewContentObjectFromSource(content, destination_slug, errors)
+    if content_id := generateNewContentRecordFromSource(content, new_instance_id, errors):
+        content_ids.append(content_id)
+
+
+def updateEntryDate(source_slug, destination_slug):
+    entry = models.Entry.objects.get(name=source_slug)
+    moveImagesOutOfADeleteFolder(entry)
+
+    new_entry, errors = createNewEntry(entry, destination_slug)
+    if errors:
+        return new_entry, errors
+
+    errors = ErrorDict()
     content_ids = []
-    # ToDo - Refactor this mess
     for content in entry.content.all():
-        content_type = content.content_type
-        
-        Model = CONTENT_MODELS[content_type]
-        obj = Model.objects.get(id=content.content_id)
-        Form = CONTENT_FORMS[content_type]
-        
-        new_obj_dict = model_to_dict(obj)
-        new_obj_dict['entry'] = destination_slug
-        
-        if content_type == "image":
-            new_obj_dict["file_path"] = Path(new_obj_dict["file_path"]).name
-        
-        new_obj_form = Form(new_obj_dict)
-        if new_obj_form.is_valid():
-            new_obj_form.save(commit=True)
-            obj.delete()
-        else:
-            errors[f'{content_type}-{obj.id}'] = new_obj_form.errors
-            
-        new_content_form = forms.ContentForm({
-            'content_type': content_type,
-            'content_id': new_obj_form.instance.pk
-        })
+        updateNewContentIds(content, destination_slug, content_ids, errors)
 
-        if new_content_form.is_valid():
-            new_content_form.save(commit=True)
-            content_ids.append(new_content_form.instance.pk)
-            content.delete()
-        else:
-            errors[f'content-{content.id}'] = new_content_form.errors
-
-    new_entry = models.Entry.objects.get(name=destination_slug)
     new_entry.content.set(content_ids)
     new_entry.save()
     if not errors:
         entry.delete()
-    
+
     return new_entry, errors
